Refactor objective card UI and improve test reliability
Extracted objective card UI and logic into a reusable PlanningIntervalObjectiveCard component, simplifying objective-list-item.tsx. Updated tests to use async act for more reliable state handling.
